paraFoam: Added exception handling to the reading of fields
Fields for which the reading fails in paraFoam are now ignored rather than
paraFoam crashing.  This change is particularly helpful while tuning the mesh
generation process allowing the mesh to be viewed even if it is inconsistent
with the field files present in the 0 directory without having to switch-off
field reading in paraFoam before loading the mesh.