meshSearch, sampledSet: Optimisation
The algorithms associated with searching for cells and faces within the
mesh have been centralised and the cached data associated with them has
been put into mesh objects. This ensures that data which is expensive to
compute and store is not duplicated and that it is correctly updated as
a result of mesh changes.

Sampled sets have been changed to ensure that they only search for their
sampling locations on demand; typically whenever a function executes
and/or writes; e.g., at a write time. This prevents unnecessary
computation in updating the sampling locations, search data, etc...,
every timestep when that data is not being used.